print.lp {lpSolve}	R Documentation

Print an lp object

Description

Print method for lp objects

Usage

## S3 method for class 'lp'
 print(x, ...)

Arguments

x
List with items named objval and status. Normally this will have been called by lp, lp.assign, or lp.transport.

...
Other arguments, all currently ignored

Details

This function prints the objective function value, together with the word "Success" if the operation is successful, or an indication of the error if not. If multiple solutions have been produced (because this was an all-binary problem and lp was called with num.bin.solns > 1) the number of solutions is also displayed.

Value

None

Author(s)

Sam Buttrey, buttrey@nps.edu

See Also

lp, lp.assign, lp.transport

[Package lpSolve version 5.6.23 Index]